+++ src/types/user.ts
@@ -1,0 +1,6 @@
+export interface UserDoc {
+  _id: string;
+  email: string;
+  name: string;
+  updated: Date;
+}
